% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CombineIndices.R
\name{CombineIndices}
\alias{CombineIndices}
\title{CombineIndices}
\usage{
CombineIndices(alpha = NULL, spp = NULL, group_name = NULL)
}
\arguments{
\item{alpha}{4-letter alpha code for single species of interest}

\item{spp}{Vector containing the 4-letter alpha codes for species group of interest}
}
\value{
Data frame containing the annual indices
}
\description{
Estimate summary indices for single species or group of species
}
