% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetBCR.R
\name{GetBCR}
\alias{GetBCR}
\title{GetBCR}
\usage{
GetBCR(alpha)
}
\arguments{
\item{alpha}{alpha code for species of interest}
}
\value{
Either a vector containing the BCR #s of interest or NULL if all BCRs where the species was detected should be used
}
\description{
Get BCR #'s to subset count data
}
