% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{MapPsi}
\alias{MapPsi}
\title{MapPsi}
\usage{
MapPsi(alpha, proj = TRUE)
}
\arguments{
\item{alpha}{4-letter alpha code for species of interest}

\item{proj}{Should maps be projected? (Probably but much slower)}
}
\description{
Annual occupancy probability maps
}
