% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OccSummary.R
\name{OccSummary}
\alias{OccSummary}
\title{OccSummary}
\usage{
OccSummary(alpha)
}
\arguments{
\item{alpha}{Vector of alpha codes for species of interest}
}
\value{
Data frame containing annual estimates of the following indices:
}
\description{
Estimate mean & CI psi for all raster cells
}
