% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{PlotLat}
\alias{PlotLat}
\title{PlotLat}
\usage{
PlotLat(alpha, ci = FALSE)
}
\arguments{
\item{alpha}{4-letter alpha code for species of interest}
}
\description{
Plot graph of changes in latitude indices
}
