% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{PlotLon}
\alias{PlotLon}
\title{PlotLon}
\usage{
PlotLon(alpha, ci = FALSE)
}
\arguments{
\item{alpha}{4-letter alpha code for species of interest}
}
\description{
Plot graph of changes in longitude indices
}
