% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{PlotRoutes}
\alias{PlotRoutes}
\title{PlotRoutes}
\usage{
PlotRoutes(alpha)
}
\arguments{
\item{alpha}{4-letter alpha code for species of interest}

\item{psi}{Should occupancy probability (in year 1) be plotted? (*doesn't work yet*)}
}
\description{
Plot map with BBS route locations
}
