% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_static.R
\name{run_static}
\alias{run_static}
\title{run_static}
\usage{
run_static(sim_data, nC = 3, nI = 25000, nB = 5000, nT = 10,
  Parallel = TRUE)
}
\arguments{
\item{sim_data}{Data object generated from sim_data_static()}

\item{nC}{Number of chains (default = 3)}

\item{nI}{Number of interations (default = 25000)}

\item{nB}{Number of adaption and burn-in iterations (default = 5000)}

\item{nT}{Thinning rate (default = 10)}

\item{Parallel}{Should chains be run in parallel (default = TRUE)}
}
\description{
Run single-season correlated detection model
}

