% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunMod.R
\name{RunMod}
\alias{RunMod}
\title{RunMod}
\usage{
RunMod(alpha, nI = 10000, nA = 5000, nC = 3, nT = 20, Parallel = TRUE)
}
\arguments{
\item{alpha}{alpha code for species of interest}
}
\description{
Run model
}
