% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data_static.R
\name{sim_data_static}
\alias{sim_data_static}
\title{sim_data_static}
\usage{
sim_data_static(nRoutes = 150, nStops = 50, nPred = 5, alpha0 = -0.5,
  alpha1 = 0.75, sigma1 = 0.5, sigma2 = 0.5, sigma3 = 0.05,
  theta = c(0.3, 0.75), sx = 0.2, sy = 0.2, nSum = NULL)
}
\arguments{
\item{nRoutes}{Number of sample units}

\item{nStops}{Number of stops within each sample unit}

\item{alpha0}{Intercept for detection model}

\item{alpha1}{Linear effect of Xp on detection probability}

\item{sigma1}{SD of Xp}

\item{sigma2}{SD of X}

\item{sigma3}{Noise}

\item{theta}{Vector containing the correlation terms}

\item{sx}{SD of lon}

\item{nSum}{Optional number of stops to aggregate}

\item{xy}{SD of lat}
}
\description{
Simulation of single-season count data for the static correlated-detection N-mixture model
}
