% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_probs.R
\name{y_probs}
\alias{y_probs}
\title{y_probs}
\usage{
y_probs(psi, xpsi, p)
}
\arguments{
\item{psi}{Predicted route-level occupancy}

\item{xpsi}{spatial correlation parameters}

\item{p}{detection probability}
}
\description{
Detection history probabilities, given model parameters
}
