% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunMod.R
\name{RunMod}
\alias{RunMod}
\title{RunMod}
\usage{
RunMod(spp = NULL, alpha = NULL, nI = 2000, nA = 1000, nC = 3,
  nT = 5, Parallel = FALSE)
}
\arguments{
\item{alpha}{alpha code for species of interest; if NULL, runs model in parallel for all species}
}
\description{
Run model
}
