% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data_static.R
\name{sim_data_static}
\alias{sim_data_static}
\title{sim_data_static}
\usage{
sim_data_static(nRoutes = 150, nStops = 50, alpha0 = -0.5,
  alpha1 = 0.75, sigma1 = 0.5, beta0 = -2, beta1 = 0.8, beta2 = -1,
  sigma2 = 0.25, sigma3 = 0.05, theta = c(0.3, 0.75), sx = 0.2,
  sy = 0.2, nSum = NULL)
}
\arguments{
\item{nRoutes}{Number of sample units}

\item{nStops}{Number of stops within each sample unit}

\item{alpha0}{Intercept for detection model}

\item{alpha1}{Linear effect of Xp on detection probability}

\item{sigma1}{SD of Xp}

\item{beta0}{Logit expected occupancy probability}

\item{beta1}{Linear effect of X on psi}

\item{beta2}{Quadratic effect of X on psi}

\item{sigma2}{SD of X}

\item{sigma3}{Noise}

\item{theta}{Vector containing the correlation terms}

\item{nSum}{Optional number of stops to aggregate}
}
\description{
Simulation of single-season count data for the static correlated-detection N-mixture model
}

